Clean up old ElmString stuff

+++ src/test/test.c
@@ -93,7 +93,7 @@ void* expect_equal(char* expect_description, void* left, void* right) {
 }
 
 ElmString* create_string(char* c_string) {
-  size_t len = (size_t)strlen(c_string);
+  size_t len = strlen(c_string);
   ElmString* s = newElmString(len);
   for (size_t i = 0; i < len; i++) {
     s->words16[i] = (u16)c_string[i];
